+++ packages/terra-slide-panel/src/SlidePanel.jsx
@@ -85,6 +85,36 @@ const defaultProps = {
   panelSize: 'small',
 };
 
+/**
+ * Checks whether or not the given node should be focusable.
+ * If not focusable, find the closest parent element that is.
+ * If it cannot find a focusable element, return undefined.
+ */
+const findFocusableElement = (node) => {
+  const focusableElementSelector = 'a[href]:not([tabindex=\'-1\']), area[href]:not([tabindex=\'-1\']), input:not([disabled]):not([tabindex=\'-1\']), '
+  + 'select:not([disabled]):not([tabindex=\'-1\']), textarea:not([disabled]):not([tabindex=\'-1\']), button:not([disabled]):not([tabindex=\'-1\']), '
+  + 'iframe:not([tabindex=\'-1\']), [tabindex]:not([tabindex=\'-1\']), [contentEditable=true]:not([tabindex=\'-1\'])';
+
+  const focusableElements = [...document.body.querySelectorAll(`${focusableElementSelector}`)].filter(
+    element => !element.hasAttribute('disabled')
+    && !element.getAttribute('aria-hidden')
+    && !!(element.offsetWidth || element.offsetHeight || element.getClientRects().length)
+    && window.getComputedStyle(element).visibility !== 'hidden'
+    && element.closest('[inert]') === null,
+  );
+
+  let currentNode = node;
+  while (currentNode.parentElement) {
+    if (focusableElements.includes(currentNode)) {
+      return currentNode;
+    }
+
+    currentNode = currentNode.parentElement;
+  }
+
+  return undefined;
+};
+
 class SlidePanel extends React.Component {
   constructor(props) {
     super(props);
@@ -97,7 +127,8 @@ class SlidePanel extends React.Component {
   componentDidUpdate(prevProps) {
     if (this.props.isOpen && this.props.isOpen !== prevProps.isOpen) {
       // Save the disclosing node for returning focus when panel is closed
-      this.setDisclosingNode(document.activeElement);
+      const disclosingNode = findFocusableElement(document.activeElement);
+      this.setDisclosingNode(disclosingNode);
       this.panelNode.focus();
       return;
     }