+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-slide-panel` example to use `terra-buttons`.
+
 ## 1.86.0 - (May 8, 2024)
 
 * Changed